iran-watch-openclaw/requirements.txt
httpx==0.27.2
feedparser==6.0.11
PyYAML==6.0.2
python-dateutil==2.9.0.post0
rapidfuzz==3.10.1
tenacity==9.0.0
langdetect==1.0.9
pydantic==2.9.2
pytest==8.3.3
